+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/Models/Responses/Package/PackageWorkItemLookupResponse.cs
@@ -2,16 +2,25 @@
 // Licensed under the MIT License.
 using System.Text;
 using System.Text.Json.Serialization;
+using Azure.Sdk.Tools.Cli.Configuration;
 
 namespace Azure.Sdk.Tools.Cli.Models.Responses.Package;
 
 public class PackageWorkItemLookupResponse : CommandResponse
 {
+    private string workItemUrl = string.Empty;
+
     [JsonPropertyName("work_item_id")]
     public int WorkItemId { get; set; }
 
     [JsonPropertyName("work_item_url")]
-    public string WorkItemUrl { get; set; } = string.Empty;
+    public string WorkItemUrl
+    {
+        get => WorkItemId > 0
+            ? $"{Constants.AZURE_SDK_DEVOPS_BASE_URL}/{Constants.AZURE_SDK_DEVOPS_RELEASE_PROJECT}/_workitems/edit/{WorkItemId}"
+            : workItemUrl.Replace("_apis/wit/workItems", "_workitems/edit");
+        set => workItemUrl = value;
+    }
 
     [JsonPropertyName("package_name")]
     public string PackageName { get; set; } = string.Empty;